Add starting code for dashboard page (#55)
* Change the accepted CORS origins

Allow only the frontend origin for cross-origin resource requests.
Previously, any domain was allowed. This change is made to allow the
frontend to successfully pass authentication credentials in an HTTP
request header.

* Allow retrievel of a given user's characters

Fulfills a new PostgreSQL character interface which describes the
ability to retrieve all characters belonging to a specific user.

* Add HTTP route for user-specific character fetch

Add a new route '/auth/character/me' which will return all the
characters belonging to a user, provided the user is authenticated using
a valid JSON web tokeAdd HTTP route for user-specific character fetch

* Remove custom axios config

Add custom configuration to the default exported axios object.

Add an HTTP request interceptor which appends an Authorization header to
outgoing requests.

* Remove axios header changes in authen store

Deleted code which incorrectly added an Authorization header to the
axios default config.

* Add new dashboard page

Add route to a new '/dashboard' page.

This dashboard page is intended to give a user the option to view
spells, items, and campaigns related to a character belonging to them.
A user selects a character and the information is fetched.

* Update nav drawer content and theme

Change colours from grayscale to blue and white.

Add functional links to other pages.