Update web theme and add spell display to the GUI (#58)
* Reorganize main.js

* Update website theme

Modify primary, secondary, and accent colours so the theme no longer
induces bleeding eyes.

* Update registration + login icon colour.

Changed to match primary theme colour.

* Provide functionality to display spells

Update the info pane to dynamically fetch data depending on which tab
was clicked by the user.

Add new spell component which will be rendered when the spell tab is
clicked by the user. So far, it displays spells in a grid/'to-do list'
style.