Add global notification snackbar (#94)
* Create notification Vuex store

Implement a new namespace store 'notifications' which stores the
message, color, and timeout state for use in a global v-snackbar
notifier component.

* Add v-snackbar to App.vue for global notifications

This snackbar component will listen for changes to the notification
store state and re-render the snackbar when the changes occur.

* Add notification for character management

Add notifications for both character creation and deletion. These
notifications will show in both successful and error states.

* Remove erroneous function call in CreationStepper

There seemed to be a call to a non-existent function
`fetchUserCharacters` which was leftover from copy-paste. This is now
removed.